docs: Pet API 문서화
