+++ animory/src/main/java/com/daggle/animory/domain/pet/service/PetReadService.java
@@ -2,7 +2,6 @@ package com.daggle.animory.domain.pet.service;
 
 
 import com.daggle.animory.common.error.exception.NotFound404;
-import com.daggle.animory.common.error.exception.UnAuthorized401;
 import com.daggle.animory.domain.account.entity.Account;
 import com.daggle.animory.domain.fileserver.S3FileRepository;
 import com.daggle.animory.domain.pet.dto.request.PetRegisterRequestDto;
@@ -42,12 +41,12 @@ public class PetService {
 
         // DTO에 넣어주기
         final List<SosPetDto> sosList = sosProfiles.stream()
-                .map(SosPetDto::fromEntity)
-                .toList();
+            .map(SosPetDto::fromEntity)
+            .toList();
 
         final List<NewPetDto> newList = newProfiles.stream()
-                .map(NewPetDto::fromEntity)
-                .toList();
+            .map(NewPetDto::fromEntity)
+            .toList();
 
         return new PetProfilesDto(sosList, newList);
     }
@@ -63,30 +62,32 @@ public class PetService {
     public PetDto getPetDetail(final int petId) {
         // petId로 Pet, PetPolygonProfile 얻어오기
         final Pet pet = petRepository.findById(petId)
-                .orElseThrow(() -> new NotFound404("해당 동물이 존재하지 않습니다."));
+            .orElseThrow(() -> new NotFound404("해당 동물이 존재하지 않습니다."));
 
         final PetPolygonProfile petPolygonProfile = petPolygonRepository.findByPetId(petId)
-                .orElseThrow(() -> new NotFound404("등록된 다각형 프로필이 존재하지 않습니다."));
+            .orElseThrow(() -> new NotFound404("등록된 다각형 프로필이 존재하지 않습니다."));
 
         return PetDto.fromEntity(pet, petPolygonProfile);
     }
 
     @Transactional
     public RegisterPetSuccessDto registerPet(final Account account,
-            final PetRegisterRequestDto petRequestDTO, final MultipartFile image,
-            final MultipartFile video) {
+                                             final PetRegisterRequestDto petRequestDTO,
+                                             final MultipartFile image,
+                                             final MultipartFile video) {
 
         // 이미지, 숏폼 파일 저장 후 url 얻어오기
+        // TODO: 이미지, 비디오 저장 요청 둘 중 하나가 실패했을 때, 또는 파일 저장에는 성공하였으나, 이후 DB Insert 등에 실패하였을 때, 파일 모두 삭제하는 Rollback 처리
         final URL imageUrl = fileRepository.save(image);
         final URL videoUrl = fileRepository.save(video);
 
         // 펫 등록을 요청한 유저의 보호소 조회
         final Shelter shelter = shelterRepository.findByAccountId(account.getId())
-                .orElseThrow(() -> new UnAuthorized401("보호소 정보가 존재하지 않습니다."));
+            .orElseThrow(() -> new NotFound404("반려동물을 등록하고자 하는 보호소 정보가 존재하지 않습니다."));
 
         // 펫 DB 저장
         final Pet registerPet = petRepository.save(
-                petRequestDTO.toEntity(shelter, imageUrl.toString(), videoUrl.toString()));
+            petRequestDTO.toEntity(shelter, imageUrl.toString(), videoUrl.toString()));
 
         // 펫 다각형 프로필 DB 저장
         petPolygonRepository.save(petRequestDTO.petPolygonProfileDto().toEntity(registerPet));
@@ -96,15 +97,15 @@ public class PetService {
 
     @Transactional
     public UpdatePetSuccessDto updatePet(final int petId,
-            final PetUpdateRequestDto petUpdateRequestDto, final MultipartFile image,
-            final MultipartFile video) {
+                                         final PetUpdateRequestDto petUpdateRequestDto, final MultipartFile image,
+                                         final MultipartFile video) {
 
         // 펫 id로 Pet, PetPolygonProfile 얻어오기
         final Pet updatePet = petRepository.findById(petId)
-                .orElseThrow(() -> new NotFound404("등록되지 않은 펫입니다."));
+            .orElseThrow(() -> new NotFound404("등록되지 않은 펫입니다."));
 
         final PetPolygonProfile petPolygonProfile = petPolygonRepository.findByPetId(petId)
-                .orElseThrow(() -> new NotFound404("등록된 다각형 프로필이 존재하지 않습니다."));
+            .orElseThrow(() -> new NotFound404("등록된 다각형 프로필이 존재하지 않습니다."));
 
         // 이미지, 비디오 파일 업데이트
         updateFile(updatePet, image, video);
@@ -118,7 +119,7 @@ public class PetService {
 
     // 이미지, 비디오 파일 수정 요청 시 기존 파일 삭제 후 업데이트
     public void updateFile(final Pet updatePet, final MultipartFile image,
-            final MultipartFile video) {
+                           final MultipartFile video) {
         if (image != null) {
             final URL imageUrl = fileRepository.save(image);
             updatePet.updateImage(imageUrl.toString());
@@ -134,10 +135,10 @@ public class PetService {
 
         // 펫 id로 Pet, PetPolygonProfile 얻어오기
         final Pet registerPet = petRepository.findById(petId)
-                .orElseThrow(() -> new NotFound404("등록되지 않은 펫입니다."));
+            .orElseThrow(() -> new NotFound404("등록되지 않은 펫입니다."));
 
         final PetPolygonProfile petPolygonProfile = petPolygonRepository.findByPetId(petId)
-                .orElseThrow(() -> new NotFound404("등록된 다각형 프로필이 존재하지 않습니다."));
+            .orElseThrow(() -> new NotFound404("등록된 다각형 프로필이 존재하지 않습니다."));
 
         // Pet -> PetRegisterInfoDto
         return PetRegisterInfoDto.fromEntity(registerPet, petPolygonProfile);
